pyproject.toml
[project]
name = "nba-game-analyzer"
version = "0.1.0"
description = "AI-powered NBA game analyzer with play-by-play insights and player performance clustering"
requires-python = ">=3.8"
dependencies = [
    "nba_api>=1.4.1",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
    "streamlit>=1.28.0",
    "plotly>=5.15.0",
    "jupyter>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"] 
